Sprites are working and main menu is mainly done

+++ proj/src/visuals/sprite.c
@@ -43,12 +43,18 @@ int (destroy_all_sprites)() {
         printf("Error: problems occured while truing to destroy -clickCursor- sprite! \n");
         return 1;
     }*/
+
+    if(destroy_sprite(wagnerzim) != 0) {
+        printf("Error: problems occured while truing to destroy -wagnerzim- sprite! \n");
+        return 1;
+    }
+
     if(destroy_sprite(titleText) != 0) {
         printf("Error: problems occured while truing to destroy -titleText- sprite! \n");
         return 1;
     }
 
-    /*if(destroy_sprite(controlShellText) != 0) {
+    if(destroy_sprite(controlShellText) != 0) {
         printf("Error: problems occured while truing to destroy -controlShellText- sprite! \n");
         return 1;
     }
@@ -71,7 +77,12 @@ int (destroy_all_sprites)() {
     if(destroy_sprite(helpText) != 0) {
         printf("Error: problems occured while truing to destroy -helpText- sprite! \n");
         return 1;
-    }*/
+    }
+
+    if(destroy_sprite(numberZero) != 0) {
+        printf("Error: problems occured while truing to destroy -numberZero- sprite! \n");
+        return 1;
+    }
 
     return 0;
 }
@@ -92,6 +103,9 @@ int (draw_sprite)(Sprite *sp) {
 }
 
 int (load_sprites)() {
+    int initial_margin = 10;
+    int buttons_margin = vmi_p.YResolution / 20;
+
     /*normalCursor = create_sprite();
     if(normalCursor == NULL) {
         printf("Error: Problems occured while trying to load -normalCursor- sprite! \n");
@@ -104,41 +118,53 @@ int (load_sprites)() {
         return 1;
     }*/
 
-    titleText = create_sprite((xpm_map_t)XPM_TitleText, 300, 400);
+    wagnerzim = create_sprite((xpm_map_t)XPM_Wagnerzim, (initial_margin), initial_margin);
+    if(wagnerzim== NULL) {
+        printf("Error: Problems occured while trying to load -wagnerzim- sprite! \n");
+        return 1;
+    }
+
+    titleText = create_sprite((xpm_map_t)XPM_TitleText, (vmi_p.XResolution / 5 + 35), initial_margin);
     if(titleText == NULL) {
         printf("Error: Problems occured while trying to load -titleText- sprite! \n");
         return 1;
     }
 
-    /*controlShellText = create_sprite();
+    controlShellText = create_sprite((xpm_map_t)XPM_ControlShellButton, (vmi_p.XResolution / 3), (initial_margin + titleText->height + buttons_margin));
      if(controlShellText == NULL) {
         printf("Error: Problems occured while trying to load -controlShellText- sprite! \n");
         return 1;
     }
 
-    housePlantText = create_sprite();
+    housePlantText = create_sprite((xpm_map_t)XPM_HousePlantButton, (vmi_p.XResolution / 3), (initial_margin + 2*titleText->height + 2*buttons_margin));
      if(housePlantText == NULL) {
         printf("Error: Problems occured while trying to load -housePlantText- sprite! \n");
         return 1;
     }
 
-    securityCameraText = create_sprite();
+    securityCameraText = create_sprite((xpm_map_t) XPM_SecurityCameraButton, (vmi_p.XResolution / 3), (initial_margin + 3*titleText->height + 3*buttons_margin));
      if(securityCameraText == NULL) {
         printf("Error: Problems occured while trying to load -securityCameraText- sprite! \n");
         return 1;
     }
 
-    displayMessageText = create_sprite();
+    displayMessageText = create_sprite((xpm_map_t) XPM_DisplayMessageButton, (vmi_p.XResolution / 3), (initial_margin + 4*titleText->height + 4*buttons_margin));
      if(displayMessageText == NULL) {
         printf("Error: Problems occured while trying to load -displayMessageText- sprite! \n");
         return 1;
     }
 
-    helpText = create_sprite();
+    helpText = create_sprite((xpm_map_t) XPM_HelpButton, (vmi_p.XResolution / 3), (initial_margin + 5*titleText->height + 5*buttons_margin));
      if(helpText == NULL) {
         printf("Error: Problems occured while trying to load -helpText- sprite! \n");
         return 1;
-    }*/
+    }
+
+    numberZero = create_sprite((xpm_map_t) XPM_Zero, (25), (initial_margin + 3*titleText->height + 3*buttons_margin));
+     if(helpText == NULL) {
+        printf("Error: Problems occured while trying to load -helpText- sprite! \n");
+        return 1;
+    }
 
     return 0;
 }
